+++ src/main/java/com/can/service/impl/DepartmentServiceImpl.java
@@ -20,7 +20,7 @@ public class DepartmentServiceImpl implements DepartmentService {
     }
 
     @Override
-    public Department getDepartmentById(Integer id) {
+    public Department getDepartmentById(Long id) {
         return departmentMapper.getDepartmentById(id);
     }
 }
